requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
distro==1.9.0
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
groq==0.36.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
mysql-connector-python==8.3.0
mysqlclient==2.2.2
numpy==2.3.5
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
